Advance RAG/requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
PyPDF2
